fix: set initial value for reducer function
The initial value (`[]`) looks like it is placed in the wrong context, so instead of being passed to the `reduce` function as the initial value, it is passed to the `useEffect` function as the second parameter. This caused the returned `data` to be the first `response` from the `Promise.all` chain with all the other responses reduced into it.