fix(alerts-service): bind id to remove
The problem with the previous implementation was that it would look at the current value of `id` when remove was called. And the result was that the last alert would always be removed. When using `.bind` it will set the value of `id` to the current value of `id` when the alert is being created, which is always correct.