fix(use-data-query): memoize refetch function
We were recreating refetch for every render which made it difficult
to use with hooks like useEffect, since specifying refetch as a
dependency would cause the useEffect hook to run every render. By
memoizing it the function will only change identity if any of its
dependencies change (so when supplying new variables for example).